Teach lit to honor conditional directives.  The syntax is:   IF(condition(value)): If the value satisfies the condition, the line is processed by lit;  otherwise it is skipped.  A test with no unignored directives is resolved as Unsupported.
The test suite is responsible for defining conditions;  conditions are unary
functions over strings.  I've defined two conditions in the LLVM test suite,
TARGET (with values like those in TARGETS_TO_BUILD) and BINDING (with values
like those in llvm_bindings).  So for example you can write:
  IF(BINDING(ocaml)): RUN: %blah %s -o -
and the RUN line will only execute if LLVM was configured with the ocaml
bindings.

llvm-svn: 97726